Create EC Integration when creating an Application
When an IntegrationTestScenario running the Enterprise Contract
executes, the UI for the corresponding Application is populated with
various security results. This is valuable information to users.

This change makes it so when an Application is created, the EC
IntegrationTestScenario is automatically created with default values.
This lowers the entrybarrier to accessing the security results.

In the future, this auto creation may move to another RHTAP component,
e.g. a new EC controller. Or, it may be desireable to introduce UI
controls to customize its initial creation, e.g. different Tekton
bundle. It's unclear at this point what our users will prefer.  For this
reason, we make the most minimal change possible to introduce the
feature to users without investing too much in a certain solution right
now.

https://issues.redhat.com/browse/HACBS-2191

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>